requirements.txt
netaddr
mysql-connector-python
tabulate
